+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/PosttestApplication.java
@@ -7,9 +7,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @SpringBootApplication
 @EnableJpaAuditing
 public class PosttestApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(PosttestApplication.class, args);
 	}
-
 }
