+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/controllers/LotteryController.java
@@ -1,5 +1,7 @@
-package com.kbtg.bootcamp.posttest.lottery;
+package com.kbtg.bootcamp.posttest.controllers;
 
+import com.kbtg.bootcamp.posttest.models.LotteriesResponse;
+import com.kbtg.bootcamp.posttest.services.LotteryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
