+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/entities/Lottery.java
@@ -1,8 +1,7 @@
-package com.kbtg.bootcamp.posttest.lottery;
+package com.kbtg.bootcamp.posttest.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.kbtg.bootcamp.posttest.core.entity.AuditEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Min;
