+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exceptions/LotterySoldOutException.java
@@ -1,11 +1,11 @@
-package com.kbtg.bootcamp.posttest.core.exceptions;
+package com.kbtg.bootcamp.posttest.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
-public class LotteryNotFoundException extends RuntimeException {
-    public LotteryNotFoundException(String message) {
+public class LotterySoldOutException extends LotteryNotFoundException {
+    public LotterySoldOutException(String message) {
         super(message);
     }
 }
