+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/models/LotteriesResponse.java
@@ -1,4 +1,4 @@
-package com.kbtg.bootcamp.posttest.lottery;
+package com.kbtg.bootcamp.posttest.models;
 
 import lombok.Data;
