+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/models/LotteryResponse.java
@@ -1,4 +1,4 @@
-package com.kbtg.bootcamp.posttest.shared.responses;
+package com.kbtg.bootcamp.posttest.models;
 
 import lombok.Data;
