+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/models/MyLotteryResponse.java
@@ -1,4 +1,4 @@
-package com.kbtg.bootcamp.posttest.user;
+package com.kbtg.bootcamp.posttest.models;
 
 import lombok.Data;
