+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/repositories/LotteryRepository.java
@@ -1,5 +1,6 @@
-package com.kbtg.bootcamp.posttest.lottery;
+package com.kbtg.bootcamp.posttest.repositories;
 
+import com.kbtg.bootcamp.posttest.entities.Lottery;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
