+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/repositories/UserRepository.java
@@ -1,5 +1,6 @@
-package com.kbtg.bootcamp.posttest.user;
+package com.kbtg.bootcamp.posttest.repositories;
 
+import com.kbtg.bootcamp.posttest.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
