+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/repositories/UserTicketRepository.java
@@ -1,5 +1,6 @@
-package com.kbtg.bootcamp.posttest.user;
+package com.kbtg.bootcamp.posttest.repositories;
 
+import com.kbtg.bootcamp.posttest.entities.UserTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
