+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/services/UserService.java
@@ -1,0 +1,7 @@
+package com.kbtg.bootcamp.posttest.services;
+
+import com.kbtg.bootcamp.posttest.models.MyLotteryResponse;
+
+public interface UserService {
+    MyLotteryResponse listAllMyLottery(Integer userId);
+}
