--- posttest/src/main/java/com/kbtg/bootcamp/posttest/user/UserService.java
@@ -1,5 +1,0 @@
-package com.kbtg.bootcamp.posttest.user;
-
-public interface UserService {
-    MyLotteryResponse listAllMyLottery(Integer userId);
-}
